+++ NEWS.md
@@ -7,6 +7,8 @@ There's now an L (Log) and E (Edit) button next to each scheduled task to view t
 
 New presets/rules:
 * 2024/11/28 - `DisableCloudOptimizedContent`/`EnableCloudOptimizedContent`/`ViewCloudOptimizedContent` -> Turn off cloud optimized content in all Windows experiences
+* 2024/11/28 - `DisableWindowsTips`/`EnableWindowsTips`/`ViewWindowsTips` -> Prevents Windows tips from being shown to users
+
 
 ## Version 3.17 (2024/10/01)